requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
python-jose==3.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1
